app/api/pyproject.toml
[tool.poetry]
name = "kol-torah-api"
version = "0.1.0"
description = "Kol Torah API backend"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.120.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
